tdd-api/pyproject.toml
[project]
name = "tdd-api"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "", email = "me@email.home" }]
dependencies = [
    "fastapi>=0.128.4",
    "uvicorn>=0.40.0",
    "sqlalchemy>=2.0.46",
    "alembic>=1.18.3",
    "kink>=0.8.1",
    "structlog>=25.5.0",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[dependency-groups]
lint = ["mypy>=1.19.1", "ruff>=0.15.0"]
test = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-watch>=4.2.0",
    "httpx>=0.28.1",
]

[tool.ruff]
include = ["**/*.py", "**/*.pyi", "**/pyproject.toml"]

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "E",  # pycodestyle - Error
    "F",  # Pyflakes
    "I",  # isort
    "W",  # pycodestyle - Warning
    "UP", # pyupgrade
]

ignore = [
    "E501", # line-too-long
    "W191", # tab-indentation
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pdm.scripts]
watch = "uvicorn src.tdd_api.main:app --port 8889 --reload"
lint = "ruff ."
fmt = "ruff format ."
test = "pytest src tests"
ptw = "ptw ."
setup-alembic = "python -m tdd_api.scripts.setup_alembic"

[tool.mypy]
strict = true

[tool.pytest-watcher]
now = false
clear = true
delay = 0.2
runner = "pytest"
runner_args = []
patterns = ["src/, tests/, *.py"]
ignore_patterns = [".venv/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
